% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{FizzBuzz}
\alias{FizzBuzz}
\title{It fizzez and buzzez baybee}
\usage{
FizzBuzz(n)
}
\arguments{
\item{n}{a vector of integers}
}
\value{
Modified 'fizz buzz' vector
}
\description{
Takes in a vector of numbers and returns a modified vector
that has 'fizz' wherever the input integer is divisible by 3
and 'buzz' wherever divisible by 5. it'll fizz AND buzz (wow) if
the input is divisible by both
}
\examples{
FizzBuzz(c(1,2,3,4,5,6,7))
FizzBuzz(my_fav_vector)
}
