% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{Flagstaff Pulliam Airport Weather Information}
\format{
A data frame with 25449 observations with 5 columns.
\describe{
\item{DATE}{date: date of the observation in format yyyy-mm-dd}
\item{PRCP}{num: precipitation in inches}
\item{SNOW}{num: snowfall in inches}
\item{TMAX}{int: maximum temperature in degrees Fahrenheit}
\item{TMIN}{int: minimum temperature in degrees Fahrenheit}
}
}
\usage{
Flagstaff_Weather
}
\description{
Data on Weather observations from the Pulliam Airport from 1950
to 2019. Contains fields for Date, Precipitation, Snow, and
Temperature minimums and maximums
}
\keyword{datasets}
